requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-dotenv==1.0.0
whitenoise==6.6.0
